Replaced glColor4x with glColor4f, maybe it will fix some isues reported by users

+++ project/jni/sdl-1.2/src/video/android/SDL_androidvideo.h
@@ -69,7 +69,7 @@ extern void SDL_ANDROID_initFakeStdout();
 extern SDL_VideoDevice *ANDROID_CreateDevice_1_3(int devindex);
 extern void SDL_ANDROID_ProcessDeferredEvents();
 extern void SDL_ANDROID_WarpMouse(int x, int y);
-extern void SDL_ANDROID_DrawMouseCursor(int x, int y, int size, int alpha);
+extern void SDL_ANDROID_DrawMouseCursor(int x, int y, int size, float alpha);
 extern void SDL_ANDROID_DrawMouseCursorIfNeeded();
 extern void SDL_ANDROID_CallJavaTogglePlainAndroidSoftKeyboardInput();
 extern void SDL_ANDROID_CallJavaStartAccelerometerGyroscope(int start);
